Sync

+++ FundsTool/Model/Payment.cs
@@ -9,11 +9,10 @@ namespace FundsTool
     public class Payment
     {
         public string Source { get; set; }
-        public int Amount { get; set; }
+        public double Amount { get; set; }
         public string Description { get; set; }
         public DateTime PostDate { get; set; }
         public DateTime TransactionDate { get; set; }
         public string Type { get; set; }
-
     }
 }
